Decouple response processing out of transparent proxy
In order to minimize the amount of transport-specific logic in the
transparent proxy, create a response processor interface and put the new
SSE rewrite logic behind that interface. Create a no-op implementation
for HTTP streamable.
